+++ dashboard/src/vars/devine_topics.json
@@ -38,5 +38,17 @@
     "body_position": {
         "name": "/body_tracking",
         "type": "std_msgs/String"
+    },
+    "player_name": {
+        "name": "/current_player",
+        "type": "std_msgs/String"
+    },
+    "game_success": {
+        "name": "/is_guesswhat_succeed",
+        "type": "std_msgs/Bool"
+    },
+    "scoreboard": {
+        "name": "/scoreboard",
+        "type": "std_msgs/String"
     }
 }
