+++ dashboard/views/index.html
@@ -4,6 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <title>🤖 DEVINE Dashboard</title>
+
   <!-- https://bootswatch.com/darkly/ -->
   <link rel="stylesheet" href="/assets/css/bootstrap.darkly.min.css">
   <link rel="stylesheet" href="/assets/css/styles.css">
