+++ dashboard/views/scoreboard/scoreboard.html
@@ -1,1 +1,15 @@
-Scoreboard
+<h1 class="text-center font-weight-bold">DEVINE Leaderboard</h1>
+
+<table class="table table-striped table-bordered table-hover text-center">
+  <thead class="thead-light">
+    <tr>
+      <th scope="col">Position</th>
+      <th scope="col">Name</th>
+      <th scope="col">Number of games</th>
+      <th scope="col">Number of game won against IRL-1</th>
+      <th scope="col">Percentage won</th>
+    </tr>
+  </thead>
+  <tbody id="tbody-leaderboard">
+  </tbody>
+</table>
